changes in properties
- use int in Dish.WeightInGrams instead of double (for accuracy)
- rename Opening and Closing time

+++ FoodDelivery.Domain/Entities/Dish.cs
@@ -18,7 +18,7 @@ public class Dish
     /// <summary>
     /// Вес блюда в граммах
     /// </summary>
-    public double WeightInGrams { get; set; }
+    public int WeightInGrams { get; set; }
 
     /// <summary>
     /// Цена блюда в рублях
